feat: implement Mercurio provider for exchange rates; add request and response models, update Altair and Orion providers to use decimal for amount

+++ Gateway/Gateway.Application/Common/Interfaces/IProvider.cs
@@ -2,5 +2,5 @@ namespace Gateway.Application.Common.Interfaces;
 
 internal interface IProvider
 {
-    Task<decimal> GetRateAsync(string fromCurrency, string toCurrency, double amount);
+    Task<decimal> GetRateAsync(string fromCurrency, string toCurrency, decimal amount);
 }
