+++ Gateway/Gateway.Application/Exchanges/Dto/ExchangeQuery.cs
@@ -1,3 +1,3 @@
 namespace Gateway.Application.Exchanges.Dto;
 
-public record ExchangeQuery(string From, string To, double Amount);
+public record ExchangeQuery(string From, string To, decimal Amount);
